refactor(): - renamed xxxMakers to xxxFactories - fixed issue with 2d vs 3d vertex creation - various minor tweaks
